Create tests/test_type_hints.py
