+++ Makefile
@@ -126,8 +126,8 @@ test-unicode:  ## Test APIs dealing with strings.
 test-contracts:  ## APIs sanity tests.
 	$(RUN_TEST) tests/test_contracts.py $(ARGS)
 
-test-type-hints:  ## Test psutil.net_connections() and Process.net_connections().
-	$(RUN_TEST) -k "TypeHints" $(ARGS)
+test-type-hints:  ## Test type hints
+	$(RUN_TEST) tests/test_type_hints.py $(ARGS)
 
 test-connections:  ## Test psutil.net_connections() and Process.net_connections().
 	$(RUN_TEST) -k "test_connections.py or net_" $(ARGS)
